Fixed build failure
This changes setuptools to not be upgraded.

The build was failing because of incompatabilities between the python
libraries packaging and setuptools.

The packaging library is installed via an OS package, and so it cannot
be upgraded by pip (uninstalled and then installed). The error given by
pip is that it is a 'distutils installed project' and cannot accurately
determine which files belong to it.

CASMHMS-6243